backend finished

+++ api/controllers/itemController.js
@@ -1,6 +1,5 @@
 var db = require('../models');
 
-
 exports.addItem = async (item_body,list_id,user_id) => {
   return db.Item.create(item_body).then(async newItem => {
     var i = await db.Item.findOneAndUpdate(
@@ -25,12 +24,16 @@ exports.updateItem = (item_id,item_body,user_id) => {
     }).populate('owner', '_id');
 };
 
-
 exports.removeItem = async (item_id,list_id,user_id) => {
-  await db.List.findOneAndUpdate(
-    {'content._id': {$lte: list_id}},
+  var l = await db.List.findOneAndUpdate(
+    {_id: list_id, 'content': {$lte: item_id}},
     {$pull: {content: item_id}},
-    { useFindAndModify: false });
-  await db.Item.findOneAndDelete({_id: item_id});
-  return 'item deleted';
+    { new: true, useFindAndModify: false });
+  try {
+    await db.Item.findOneAndDelete({_id: item_id, owner: user_id});
+  } catch (err) {
+    return (err.toString());
+  }
+  if (l == null) { l = 'item not found'}
+  return l;
 };
